# m8flow-backend/src/m8flow_backend/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: m8flow-backend-extension
paths:
...
  /process-models:
    get:
      summary: Processes page — enriched process model list
      description: >-
        Designer Processes models table: identity, display names, group,
        last_run_in_seconds, runs_30d. Concrete tenant required (cookie
        and/or tenantId for super-admins); never merges catalogs across
        tenants. Callers without process-model list permission get an
        empty list (200), not a 403. Optional group filters to models
        under that process-group path (unknown group → []). Does not
        replace thin GET /v1.0/process-models.
      operationId: m8flow_backend.routes.processes_controller.list_process_models
      tags:
        - Processes
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
          description: >-
            Super-admin-only concrete tenant. Required for super-admins
            when m8flow_selected_tenant is unset; ignored for non-super-admins.
        - name: group
          in: query
          schema:
            type: string
          description: Optional process-group path filter.
      responses:
        "200":
          description: Process models retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessModelListResponse"
        "400":
          description: Missing concrete tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
    post:
      summary: Processes page — create a process model
      description: >-
        Creates a process model under an existing process group and writes a
        default startable BPMN (`{leaf}.bpmn`). `id` is optional: when omitted,
        the leaf is slugified from `display_name` (same as the previous create
        form). Super-admin is always 403. Distinct from thin POST
        /v1.0/process-models `{path, xml}`.
      operationId: m8flow_backend.routes.processes_controller.create_process_model
      tags:
        - Processes
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
          description: Super-admin-only concrete tenant override (writes still 403).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessModelCreateRequest"
      responses:
        "201":
          description: Process model created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessModelIdentityResponse"
        "400":
          description: Missing concrete tenant, group, id, or display name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
        "403":
          description: Not permitted to create a process model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Process group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Process model already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /process-instances:
    get:
      summary: Process Instances page — list
      description: >-
        Designer Process Instances list: id, process model identity,
        status, started_by, start/end timestamps. Concrete tenant
        required (same posture as GET /process-models — no merged
        all-tenant catalog). Optional status/search filters, real
        pagination. Callers without process-instance list permission get
        an empty page (200), not a 403.
      operationId: m8flow_backend.routes.process_instances_controller.list_process_instances
      tags:
        - ProcessInstances
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
          description: >-
            Super-admin-only concrete tenant. Required for super-admins
            when m8flow_selected_tenant is unset; ignored for non-super-admins.
        - name: status
          in: query
          schema:
            type: string
          description: Exact status filter (e.g. running, complete, error, waiting).
        - name: search
          in: query
          schema:
            type: string
          description: Case-insensitive substring match on process model identifier/display name.
        - name: started_by
          in: query
          schema:
            type: string
          description: >-
            Exact initiator-username match. Options come from
            GET /process-instances/owners.
        - name: sort
          in: query
          schema:
            type: string
            default: newest
          description: >-
            Result ordering: newest, oldest, recent_start, or status. Unknown
            values fall back to newest-first (the controller degrades gracefully,
            so no enum constraint is enforced here — ticket 03 owns any decision
            to tighten this into a validated enum).
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 25
            maximum: 100
      responses:
        "200":
          description: Process instances retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessInstanceListResponse"
        "400":
          description: Missing concrete tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
...
components:
  schemas:
...
    ProcessModelListItem:
      type: object
      required:
        - id
        - display_name
        - group_id
        - group_display_name
        - runs_30d
      properties:
        id:
          type: string
          description: Process model path relative to the tenant catalog root
        display_name:
          type: string
        group_id:
          type: string
          description: Parent process-group path; empty when the model is at tenant root
        group_display_name:
          type: string
        last_run_in_seconds:
          type: integer
          nullable: true
          description: Epoch seconds of the most recent instance start; null if never run
        runs_30d:
          type: integer
          description: Instance starts in the last 30 days
    ProcessModelListResponse:
      type: array
      items:
        $ref: "#/components/schemas/ProcessModelListItem"
...
    ProcessModelCreateRequest:
      type: object
      required:
        - group_id
      properties:
        id:
          type: string
          description: >-
            Leaf identifier (single path segment). Optional when
            display_name is set; slugified from display_name otherwise.
        group_id:
          type: string
          description: Existing process group path
        display_name:
          type: string
          description: >-
            Human-readable name. Required when id is omitted; used to
            generate the leaf identifier.
        description:
          type: string
...
    ProcessModelIdentityResponse:
      type: object
      required:
        - id
        - display_name
        - description
        - group_id
        - group_display_name
      properties:
        id:
          type: string
        display_name:
          type: string
        description:
          type: string
        group_id:
          type: string
        group_display_name:
          type: string
...
    ProcessInstancePagination:
      type: object
      required:
        - count
        - total
        - pages
      properties:
        count:
          type: integer
          description: Number of rows in this page's results.
        total:
          type: integer
          description: Total rows matching the filters, across all pages.
        pages:
          type: integer
    ProcessInstanceListItem:
      type: object
      required:
        - id
        - process_model_identifier
        - process_model_display_name
        - status
        - started_by
      properties:
        id:
          type: integer
        process_model_identifier:
          type: string
        process_model_display_name:
          type: string
        status:
          type: string
        started_by:
          type: string
        start_in_seconds:
          type: integer
          nullable: true
        end_in_seconds:
          type: integer
          nullable: true
    ProcessInstanceListResponse:
      type: object
      required:
        - results
        - pagination
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/ProcessInstanceListItem"
        pagination:
          $ref: "#/components/schemas/ProcessInstancePagination"
...
    ErrorData:
      type: object
      properties:
        error_code:
          type: string
        message:
          type: string
    ErrorResponse:
      $ref: "#/components/schemas/ErrorData"
...